+++ ui/components/datetime-picker/base/example.jsx
@@ -11,7 +11,7 @@ import {
 import { DatePicker } from '../../datepickers/base/example';
 import ButtonIcon from '../../button-icons/';
 import { FormElement } from '../../form-element';
-import { Input } from '../../input/base/example';
+import Input from '../../input/';
 
 /* -----------------------------------------------------------------------------
     Variables